% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{categ}
\alias{categ}
\title{Simulated bird SAD dataset with species classification data}
\format{A dataframe with three columns: 1) 'abundances' = the abundance of
  each species, 2) 'species' = the species names, and 3) 'status' the species
  origin classification. In regards to (3) each species is classified as
  either native (N), exotic (E) or invasive (I).}
\source{
This package.
}
\description{
A randomly generated bird SAD dataset where each species has
  been randomly classified according to its origin (native, exotic or
  invasive).
}
\examples{
data(categ, package = "gambin")
}
