% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgambin.R, R/gambin_exp.R
\name{dgambin}
\alias{dgambin}
\alias{gambin_exp}
\title{Calculate the gambin distribution}
\usage{
dgambin(x, alpha, maxoctave, log = FALSE)

gambin_exp(alpha, maxoctave, total_species)
}
\arguments{
\item{x}{vector of (non-negative integer) quantiles.}

\item{alpha}{The shape parameter of the GamBin distribution.}

\item{maxoctave}{The scale parameter of the GamBin distribution - which octave is the highest in the empirical dataset?}

\item{log}{logical; Tf \code{TRUE}, probabilities p are given as log(p).}

\item{total_species}{The total number of species in the empirical dataset}
}
\value{
A vector with length MaxOctave + 1 of the expected number of species in each octave
}
\description{
Calculate the expected number of species in octaves for a given value of alpha and maxoctave
}
\details{
\code{dgambin} gives the distribution function of gambin, so all octaves sum to 1.
\code{gambin_exp} multiplies this by the total number of species to give the expected GamBin distribution in units of species, 
for comparison with empirical data.
}
\examples{
expected <- gambin_exp(4, 13, 200)
plot(expected, type = "l")
}
\references{
Ugland, K.I., Lambshead, F.J.D., McGill, B.J., Gray, J.S., O'Dea, N., Ladle, R.J. & Whittaker, R.J. (2007). Modelling dimensionality in species abundance distributions: description and evaluation of the Gambin model. Evolutionary Ecology Research, 9, 313-324.
Matthews, T.J., Borregaard, M.K., Ugland, K., Borges, P.A.V., Rigal, F. & Whittaker, R.J. (Early View Online). 
The gambin model provides a superior fit to species abundance distributions with a single free parameter: evidence, 
implementation and interpretation. Ecography, DOI: 10.1111/ecog.00861
}

