\name{moths}
\alias{moths}
\docType{data}
\title{
Williams' Rothamsted moth data
}
\description{
Macro-Lepidoptera captured in a light trap at Rothamsted Experimental Station during 1935
}
\usage{data(moths)}
\format{
	A numerical vector with the abundance of 195 moth species.

 
}

\source{
Williams, C.B. (1964) Patterns in the balance of nature. Academic Press, London.
}

\examples{
data(moths)
}
\keyword{datasets}
