% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gambin.R
\name{fitGambin}
\alias{fitGambin}
\title{Fit the GamBin model to a species abundance distribution}
\usage{
fitGambin(abundances, subsample = 0)
}
\arguments{
\item{abundances}{Either a vector of abundances of all species in the sample/community; or the result of \code{create_octaves}}

\item{subsample}{The number of individuals to sample from the community before fitting the GamBin model. If subsample == 0 the entire community is used}

\item{object}{a \code{gambin} object created by \code{fitGambin}}

\item{x}{A \code{gambin} object created by \code{fitGambin}}

\item{parm}{The parameter to calculate confidence intervals from. Only alpha is implemented}

\item{level}{The significance level of the confidence intervals}

\item{barcol}{The colour of the bars illustrating the empirical abundance of species in octaves}

\item{barwidth}{The width of the bars illustrating the empirical abundance of species in octaves}

\item{cex.dots}{The size of the dots illustrating the fitted abundance of species in octaves}

\item{dotpch}{The point character of the dots illustrating the fitted abundance of species in octaves}

\item{dotcol}{The colour of the dots illustrating the fitted abundance of species in octaves}

\item{line}{Should the dots be connected with a line?}

\item{lwd}{The width of the line connecting dots}

\item{linecol}{The colour of the line connecting dots}

\item{\dots}{Further arguments to pass to \code{barplot}}
}
\value{
The \code{fitGambin} function returns an object of class \code{gambin}, with the \code{alpha} 
and \code{MaxOctave} parameters of the GamBin distribution, the likelihood of the fit, and the empirical distribution over octaves.
}
\description{
Uses maximum likelihood methods to fit the GamBin model to binned 
species abundances. To control for the effect of sample size, the abundances 
may be subsampled prior to fitting.
}
\details{
The gambin distribution is fit to the number of species in abundance octaves, 
as specified by the \code{create_octaves} function. Because the shape of species abundance 
distributions depend on sample size, abundances of different communities should be compared 
on equally large samples. The sample size can be set by the \code{subsample} parameter. 
To estimate \code{alpha} from a standardised sample, the function must be run several 
times; see the examples. The \code{plot} method creates a barplot showing the observed 
number of species in octaves, with the fitted GamBin distribution shown as black dots.
}
\examples{
data(moths)
fit <- fitGambin(moths) 
plot(fit)
summary(fit)
# gambin parameters based on a standardized sample size of 1000 individuals
stand_fit <- replicate(20, fitGambin(moths, 1000)$Alpha) #may take a while on slower computers
print(c(mean = mean(stand_fit), sd = sd(stand_fit)))
}

