% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{fly}
\alias{fly}
\title{Brazilian Horse Fly Data}
\format{A list with two elements. The first element contains a numerical
  vector with the abundance of 164 fly species sampled at various sites
  across Brazil. The second element contains a numerical vector with the
  abundance of 58 fly specie sampled at a single site within Brazil using
  just canopy traps.}
\source{
This package.
}
\description{
Horse flies captured using various sampling methods at different sites
  across Brazil.
}
\examples{
data(fly, package = "gambin")
}
