% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{moths}
\alias{moths}
\title{Williams' Rothamsted moth data}
\format{A numerical vector with the abundance of 195 moth species.}
\source{
Williams, C.B. (1964) Patterns in the balance of nature. Academic Press, London.
}
\description{
Macro-Lepidoptera captured in a light trap at Rothamsted Experimental Station during 1935.
}
